tests/test_requirements.txt
pytest==7.3.1
pytest-asyncio==0.21.0
aiohttp==3.8.4
python_dotenv==1.0.0